+++ src/components/LayerList/Legend.js
@@ -87,7 +87,7 @@ const buildStyle = lyr => {
             key={i}
             sx={{
               margin: 0,
-              marginLeft: '15%',
+              marginLeft: 4,
               padding: 0,
               display: 'flex'
             }}
@@ -130,7 +130,7 @@ const buildStyle = lyr => {
   // circle-color, circle-opacity, circle-stroke-color, circle-stroke-opacity, circle-radius
   function circle(paint) {
     // circle-color, circle-stroke-color, circle-radius
-    let cc, /*csc,*/ cr;
+    let cc, /*csc,*/ cr, flexGrow, textAlign;
 
     // Is circle-color a data driven paint style?
     cc = paint.get('circle-color').value.value.toString();
@@ -153,8 +153,12 @@ const buildStyle = lyr => {
       paint.get('circle-radius').value._parameters &&
       paint.get('circle-radius').value._parameters.hasOwnProperty('stops')
     ) {
+      flexGrow = 6;
+      textAlign = 'center';
       cr = paint.get('circle-radius').value._parameters.stops;
     } else {
+      flexGrow = 1;
+      textAlign = 'left';
       cr = [paint.get('circle-radius').value.value.toString()];
     }
 
@@ -175,12 +179,12 @@ const buildStyle = lyr => {
             key={i}
             sx={{
               margin: 0,
-              marginLeft: '15%',
+              marginLeft: 4,
               padding: 0,
               display: 'flex'
             }}
           >
-            <Box sx={{ textAlign: 'center' }}>
+            <Box sx={{ flexGrow: flexGrow, textAlign: textAlign }}>
               <svg width={cr} height={cr}>
                 <circle
                   cx={cr / 2.0}
@@ -194,6 +198,7 @@ const buildStyle = lyr => {
             </Box>
             <Box
               sx={{
+                flexGrow: 1,
                 lineHeight: s[1] >= 10 ? `${+s[1] * 1.8388 - 7.5}px` : '15px'
               }}
             >
@@ -220,20 +225,89 @@ const buildStyle = lyr => {
 
   // line-cap, line-join, line-opacity, line-color, line-width, line-dasharray, line-gradient
   function line(paint) {
-    if (paint.get('line-color').value.kind === 'composite') {
+    let lc;
+    console.log(paint.get('line-color'));
+    const kind = paint.get('line-color').value.kind;
+    if (kind === 'source' || kind === 'constant') {
+      //is it a data-driven style?
+      if (
+        paint.get('line-color').value._parameters &&
+        paint.get('line-color').value._parameters.hasOwnProperty('stops')
+      ) {
+        lc = paint.get('line-color').value._parameters.stops;
+      } else {
+        lc = [paint.get('line-color').value.value.toString()];
+      }
+
+      const Items = () => {
+        let styles = [];
+        if (lc.length > 1) {
+          styles = lc.map(l => {
+            return [...l];
+          });
+        } else {
+          styles = [[null, lc[0]]];
+        }
+
+        const svgs = styles.map((s, i) => {
+          return (
+            <ListItem
+              key={i}
+              sx={{
+                margin: 0,
+                marginLeft: 4,
+                padding: 0,
+                display: 'flex',
+                alignItems: 'bottom'
+              }}
+            >
+              <Box sx={{ textAlign: 'center' }}>
+                <svg width="25" height="25">
+                  <rect
+                    x="0"
+                    y="0"
+                    rx="0"
+                    ry="0"
+                    width="25"
+                    height="7.5"
+                    sx={{ fill: s, strokeWidth: 2 }}
+                  />
+                </svg>
+              </Box>
+              <Box>
+                <Text /* sx={{ padding: '3px' }} */>{s[0]}</Text>
+              </Box>
+            </ListItem>
+          );
+        });
+
+        return <Box>{svgs}</Box>;
+      };
+
+      return (
+        <List
+          sx={{
+            margin: 0,
+            padding: 0
+          }}
+        >
+          <Items />
+        </List>
+      );
+    } else if (paint.get('line-color').value.kind === 'composite') {
       try {
-        const lc = paint
+        lc = paint
           .get('line-color')
           .value._styleExpression.expression.outputs[0].outputs[0].value.toString();
         return (
           <List
             sx={{
               margin: 0,
-              marginLeft: '15%',
+              marginLeft: 4,
               padding: 0
             }}
           >
-            <ListItem sx={{ margin: 0, marginLeft: '15%', padding: 0 }}>
+            <ListItem sx={{ margin: 0, marginLeft: 4, padding: 0 }}>
               <svg width="25" height="25">
                 <rect
                   x="0"